requirements.txt
annotated-types==0.7.0
astroid==3.3.11
colorama==0.4.6
dill==0.4.0
greenlet==3.2.4
isort==6.0.1
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
pylint==3.3.8
python-dotenv==1.1.1
SQLAlchemy==2.0.43
sqlmodel==0.0.24
tomlkit==0.13.3
typing-inspection==0.4.1
typing_extensions==4.14.1
